+++ metadata.rb
@@ -4,11 +4,12 @@ maintainer_email 'oswald@continuousphp.com'
 license          'Apache 2.0'
 description      'Installs/Configures aws-codedeploy-agent'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.3.0'
+version          '0.4.0'
 
 supports 'ubuntu'
 supports 'centos'
 supports 'fedora'
+supports 'debian'
 
 depends 'ohai'
 depends 'build-essential'