+++ CHANGELOG.md
@@ -3,14 +3,19 @@ aws-codedeploy-agent CHANGELOG
 
 This file is used to list changes made in each version of the aws-codedeploy-agent cookbook.
 
+0.4.0
+-----
+- Oswald De Riemaecker - Added Ubuntu 12.04 and Debian 8 Support
+- Oswald De Riemaecker - Adding .foodcritic
+
 0.3.0
 -----
-- Oswald De Riemaecker - Added awscli Support for onpremise support 
-- Oswald De Riemaecker - Split manual and package installer 
+- Oswald De Riemaecker - Added awscli Support for onpremise support
+- Oswald De Riemaecker - Split manual and package installer
 
 0.2.0
 -----
-- Oswald De Riemaecker - Fixed deployment root folder 
+- Oswald De Riemaecker - Fixed deployment root folder
 
 0.1.0
 -----